+++ packages/brick-container/deploy/update_postscript.sh
@@ -3,7 +3,7 @@
 # easyops 安装根目录
 install_base="/usr/local/easyops"
 
-plugin_name='brick_next'
+plugin_name='brick_next_v3'
 
 install_path="${install_base}/${plugin_name}"
 
@@ -36,8 +36,8 @@ function report_package() {
   install_base=$1
   org=$2
   install_path=$3
-  if [[ -f ${install_base}/brick_next/packages/brick-container/tools/report_installed_brick_next_package.py ]];then
-      ${install_base}/python/bin/python ${install_base}/brick_next/packages/brick-container/tools/report_installed_brick_next_package.py ${org} ${install_path}
+  if [[ -f ${install_base}/brick_next_v3/packages/brick-container/tools/report_installed_brick_next_package.py ]];then
+      ${install_base}/python/bin/python ${install_base}/brick_next_v3/packages/brick-container/tools/report_installed_brick_next_package.py ${org} ${install_path}
       if [[ $? -ne 0 ]]; then
           echo "report brick next error"
           exit 1