+++ packages/brick-container/tools/report_installed_brick_next_package.py
@@ -1,4 +1,3 @@
-
 # -*- coding: utf-8 -*-
 import logging
 import os
@@ -11,17 +10,27 @@ import tarfile
 
 from copy import deepcopy
 
+reload(sys)
+sys.setdefaultencoding("utf-8")
+
 logger = logging.getLogger("report_installed_brick_next_package")
 logging.basicConfig(level=logging.DEBUG,
                     filename="./report_installed_brick_next_package.log")
 
-
-# 1. 获取到当前的需要处理的包处理到包名
-# 2. 拿到包下面的三个文件 bricks.json stories.json contracts.json
-# 3. 读取三个文件的内容
-# 4. 调用接口，发送文件内容
 class NameServiceError(Exception):
-    pass
+  pass
+
+def join_host_port(host, port):
+    template = "%s:%s"
+    host_requires_bracketing = ':' in host or '%' in host
+    if host_requires_bracketing:
+        template = "[%s]:%s"
+    return template % (host, port)
+
+session_id, ip, port = ens_api.get_service_by_name("", "logic.micro_app_service")
+if session_id <= 0:
+    raise Exception("get name service logic.micro_app_service failed, no session_id")
+MICRO_APP_ADDR = join_host_port("127.0.0.1", port)
 
 
 def get_snippets_from_stories(stories_content):
@@ -166,16 +175,15 @@ def collect_stories(install_path):
     examples_path = os.path.join(install_path, "dist", "examples.json")
     types_path = os.path.join(install_path, "dist", "types.json")
     bricks_path = os.path.join(install_path, "dist", "bricks.json")
-    with open(bricks_path) as bricks_file:
-        bricks_content = simplejson.load(bricks_file)
-    isV3Brick = "id" in bricks_content
+    if not os.path.exists(stories_path):
+      return []
     with open(stories_path) as stories_file:
         stories_content = simplejson.load(stories_file)
     # v2 brick
-    if isV3Brick == False:
+    if not os.path.exists(manifest_path):
         return stories_content
     # v3 brick
-    elif isV3Brick:
+    elif os.path.exists(manifest_path):
         stories_list = []
         with open(manifest_path) as manifest_file:
             manifest_content = simplejson.load(manifest_file)
@@ -231,15 +239,9 @@ def collect(install_path):
 
 def report_bricks_atom(org, nb_targz_path, package_name, package_version, bricks_content, stories_content,
                        snippets_content):
-    session_id, ip, port = ens_api.get_service_by_name(
-        "web.brick_next", "logic.micro_app_service")
-    if session_id <= 0:
-        raise NameServiceError(
-            "get nameservice logic.micro_app_service error, session_id={}".format(session_id))
-    address = "{}:{}".format(ip, port)
     headers = {"org": str(org), "user": "defaultUser"}
     # report atom
-    atom_url = "http://{}/api/v1/brick/atom/import".format(address)
+    atom_url = "http://{}/api/v1/brick/atom/import".format(MICRO_APP_ADDR)
     data_dict = {"stories": stories_content, "bricks": bricks_content}
     data_str = simplejson.dumps(data_dict)
     data = {"packageName": package_name,
@@ -248,7 +250,7 @@ def report_bricks_atom(org, nb_targz_path, package_name, package_version, bricks
                         "file": open(nb_targz_path, "rb")})
     rsp.raise_for_status()
     # report snippet
-    snippet_url = "http://{}/api/v1/brick/snippet/import".format(address)
+    snippet_url = "http://{}/api/v1/brick/snippet/import".format(MICRO_APP_ADDR)
     snippet_param = {"packageName": package_name, "snippets": snippets_content}
     rsp = requests.post(snippet_url, json=snippet_param, headers=headers)
     rsp.raise_for_status()
@@ -256,15 +258,9 @@ def report_bricks_atom(org, nb_targz_path, package_name, package_version, bricks
 
 def report_brick_next_package(org, brick_targz_path, package_name, package_version):
     # report brick_next or NT
-    session_id, ip, port = ens_api.get_service_by_name(
-        "web.brick_next", "logic.micro_app_service")
-    if session_id <= 0:
-        raise NameServiceError(
-            "get nameservice logic.micro_app_service error, session_id={}".format(session_id))
-    address = "{}:{}".format(ip, port)
     headers = {"org": str(org), "user": "defaultUser"}
     # report atom
-    atom_url = "http://{}/api/v1/brick_next/report".format(address)
+    atom_url = "http://{}/api/v1/brick_next/report".format(MICRO_APP_ADDR)
     data = {"packageName": package_name, "packageVersion": package_version}
     rsp = requests.post(atom_url, data=data, headers=headers, files={
                         "file": open(brick_targz_path, "rb")})
@@ -272,16 +268,9 @@ def report_brick_next_package(org, brick_targz_path, package_name, package_versi
 
 
 def report_provider_into_contract(org, package_name, contract_content):
-    session_id, ip, port = ens_api.get_service_by_name(
-        "web.brick_next", "logic.micro_app_service")
-    if session_id <= 0:
-        raise NameServiceError(
-            "get nameservice logic.micro_app_service error, session_id={}".format(session_id))
-    address = "{}:{}".format(ip, port)
     headers = {"org": str(org), "user": "defaultUser"}
     # report contract
-    url = "http://{}/api/v1/brick/provider/import_into_contract".format(
-        address)
+    url = "http://{}/api/v1/brick/provider/import_into_contract".format(MICRO_APP_ADDR)
     param = {"packageName": package_name, "data": {
         "contractInfo": contract_content}}
     rsp = requests.post(url, json=param, headers=headers)
@@ -384,7 +373,7 @@ if __name__ == "__main__":
     if install_path.endswith(os.sep):
         install_path = install_path[:-1]
 
-    if install_path.endswith("brick_next"):
+    if install_path.endswith(("brick_next_v3", "brick_next")):
         report_brick_next(org, install_path)
     elif install_path.endswith("-NB"):
         report_nb(org, install_path)