+++ packages/preview/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.65](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.64...@next-core/preview@0.2.65) (2023-12-28)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.64](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.63...@next-core/preview@0.2.64) (2023-12-27)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.63](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.62...@next-core/preview@0.2.63) (2023-12-26)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
 ## [0.2.62](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.61...@next-core/preview@0.2.62) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/preview