+++ packages/yo/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.32](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.31...@next-core/yo@1.5.32) (2023-12-28)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.31](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.30...@next-core/yo@1.5.31) (2023-12-27)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.30](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.29...@next-core/yo@1.5.30) (2023-12-26)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
 ## [1.5.29](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.28...@next-core/yo@1.5.29) (2023-12-25)
 
 **Note:** Version bump only for package @next-core/yo